Adds a directory for Aeon branding materials
This comprises a stylesheet, header HTML, and footer HTML. They should be
self-explanatory, and work within Aeon's existing markup and styles.

Please note: these files are also subject to change by Atlas staff - so
if you ever get tasked with making an adjustment, please coordinate work
with UX. It will probably be necessary to grab the current files from
the production system, because we do not always get notified about
changes that they make - so our files may not be current.