Adds javascript and styles to show WordPress alerts
The javascript class is a more portable version of what we're using within the parent theme.

The styles are a copy-paste of the style rules from the parent theme, with a small number of adapter styles in a separate section that should handle the local customization for this markup scheme.