Copy upstream files from LibApps
Why these changes are being introduced:

At some point, certain LibApps header, footer, and
CSS/JS files were updated in the LibApps staff UI
and not committed to GitHub. We should commit these
files before making further updates.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/UXWS-1510

How this addresses that need:

This updates the repo with the latest files from
Springshare.

Side effects of this change:

* A newline was added to the end of some files. This
was IDE behavior, not an upstream change.
* Many of the changes are switching from an escaped
ampersand character to a regular ampersand character.
Escaping the ampersand is not necessary in any of
these cases, and it may make editing the files more
intuitive for other DLS staff.