Update readme
Why these changes are being introduced:

There are some details on how to update LibApps products that
would be useful to add to the docs.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/UXWS-1510 (not directly
related, but prompted by work on this ticket)

How this addresses that need:

This adds a few notes to the readme on how to access the LibApps
admin UI and navigate to the 'Look & Feel' dashboard.

Side effects of this change:

None.