Add Matomo tracking code to LibGuides
Why these changes are being introduced:

We want to collect analytics data from LibGuides in Matomo.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/UXWS-1566

How this addresses that need:

This adds Matomo tracking code to the custom head for LibGuides.
This is already in prod, and the markup in `libguides/custom-head.html`
was copied and pasted directly from our LibGuides account.

Side effects of this change:

Not a side-effect, but a note that this data is compiled in the
same property as the Libraries website.