Add upstream changes to parent theme
Why these changes are being introduced:

The basis for the `custom-head.html` is the parent theme CSS, but
we likely haven't pulled any changes from the parent theme since we
created this repo.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/PW-71

How this addresses that need:

This pulls CSS for the numbered sections in `custom-head.html`
from the compiled and minified `global.css` in WordPress.

Side effects of this change:

* Changes to the header and footer HTML will be needed to ensure
that all elements render correctly. These changes will come in a
subsequent commit to make it clearer what is being changed.
* There is, quite likely, a lot of CSS copied to `custom-head.html`
that is not necessary. It could be a worthwhile exercise to try
simplifying these files, but it is out of scope of this ticket.
* The relevant sections are not labeled in the upstream CSS. They
are compiled logically thanks to Sass partials, so it's not
difficult to locate the segments of interest, but I thought I
should mention this in case it's useful to the next person who
does this. (Another option is to copy directly from the Sass partials
themselves, but they reference modules that are further upstream,
so some modifications will be required.)