Update Springshare apps with new MIT branding
Why these changes are being introduced:

The Institute has refreshed its branding.

Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/PW-71
* https://mitlibraries.atlassian.net/browse/PW-65

How this addresses that need:

This updates the `custom-header`, `custom-footer`, and
`custom-head.html` files for all Springshare properties, excluding
LibStaff and the beta A-Z Databases. (The latter is ticketed
separately and will come in a subsequent commit.)

The all three files should be identical for LibGuides, LibAnswers,
and LibCal. The other two properties have slight differences:

* LibAnswers DDC's header has no nav. The header consists of
the MITL logo and the new MIT logo.
* LibWizard has no footer. As such, the header nav is replaced
with links to the libraries' privacy and accessibility policies.

Otherwise, the only difference should be the app name in the
'Local overrides' comment.

Side effects of this change:

* The new favicons have been added to the LibGuides `custom-head`.
(These were added to the other `custom-head` files in the
last commit, which pulled in upstream changes from WordPress.)
* The nav html has been removed from LibAnswers DDC, as it does
not have a nav and shouldn't have a hamburger menu in mobile.
* Hover/focus highlights have been fixed to extend to the full
height of the header.
* Despite loading the favicons at the top of `custom-head`,
Springshare prioritizes the Springshare favicon. It's not clear
to me how to fix this.
* To remain consistent with our previous practices, all of the
changes were made in the 'Local overrides' sections of the
`custom-head` files. We should consider a longer term solution
of trimming the `custom-heads` to what is actually needed,
and incorporating the local overrides so that we aren't repeating
media queries and rules further down in the files.