Copy LibGuides custom-head for use in beta A-Z Databases styling
Why these changes are being introduced:

Springshare will be rolling out a new A-Z Databases view soon.
They are currently providing a 'preview' mode of the beta version,
which we'd like to customize.

Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/UXWS-1621

How this addresses that need:

This copies `libguides/custom-head.html` to
`libguides/custom-head-az.html` for further customization. The file
is simply being copied as is in this commit, so it's clearer what
is being changed in the subsequent commit.

Once the new A-Z Databases launches, there will be one 'Look & Feel'
section that affects both it and LibGuides. At that point, we can
fold these changes back into `libguides/custom-head.html` and delete
the extra file.

Side effects of this change:

It's an additional file to maintain in this repo, but hopefully
not for long.