Styling updates for A-Z Databases redesign
Why these changes are being introduced:

Springshare will soon release a revamped A-Z Databases. We need
to update our custom CSS and JS to accommodate the changes.

Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/UXWS-1621

How this addresses that need:

* Adds requested changes, primarily adding the new MIT branding and
improving the styling of the form elements. The redesigned A-Z also
uses Bootstrap 5 instead of 3, so different media queries are needed
for the new breakpoints.
* Moves custom JS to `custom-footer-js` and `custom-header-js`
files, as the new System Settings allow for this. (They oddly still
require everything to be wrapped in `script` tags, which is why
these are still HTML files.)

Side effects of this change:

* Springshare has indicated that the current System Settings
(i.e., custom CSS/JS) for A-Z will eventually also apply to LibGuides.
If they do this before updating LibGuides to Bootstrap 5, then the
media queries set for A-Z won't correspond with the ones in LibGuides.
I've asked Springshare for more clarification on this, and Dhruti and
I will develop a backup plan (most likely a fluid header/footer) if
they intend to collapse the settings before updating LibGuides.
* The footer background png had been using a relative path. I updated
this to the full URL in both the LibGuides and A-Z CSS.
* Matomo code has _not_ been added to the footer JS. We currently add
it to the Custom Analytics section of  LibGuides' system settings. It's
unclear to me if this will remain  the case when the new A-Z Databases
launches, but we'll want to check on it post-launch as it may affect
data collection.